pi-changes

+++ MPU6050_changes/MPU6050.h
@@ -110,7 +110,7 @@ class MPU6050 {
 		MPU6050(int8_t addr);
 		MPU6050(int8_t addr, bool run_update_thread);
 		void getAccelRaw(float *x, float *y, float *z);
-		void getGyroRaw(float *roll, float *pitch, float *yaw);
+		void getGyroRaw(int16_t *roll, int16_t *pitch, int16_t *yaw);
 		void getAccel(float *x, float *y, float *z);
 		void getGyro(float *roll, float *pitch, float *yaw);
 		void getOffsets(float *ax_off, float *ay_off, float *az_off, float *gr_off, float *gp_off, float *gy_off);
